♻️ refacto config widget
